# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/client/repair-cases/{id}": {
      "get": {
        "operationId": "RepairCaseController_getMine",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "repairs"
        ]
      }
    },
    "/staff/repair-cases/{id}": {
      "get": {
        "operationId": "RepairCaseController_getForStaff",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "repairs"
        ]
      }
    },
    "/staff/repair-cases/{id}/status": {
      "patch": {
        "operationId": "RepairCaseController_changeStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "repairs"
        ]
      }
    },
    "/staff/repair-cases/{id}/assign/{technicianId}": {
      "post": {
        "operationId": "RepairCaseController_assign",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "technicianId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "repairs"
        ]
      }
    },
    "/staff/repair-cases/{id}/internal-notes": {
      "post": {
        "operationId": "RepairCaseController_addInternalNote",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "repairs"
        ]
      }
    },
    "/staff/repair-cases/{id}/client-messages": {
      "post": {
        "operationId": "RepairCaseController_addClientMessageFromStaff",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "repairs"
        ]
      }
    },
    "/client/repair-cases/{id}/messages": {
      "get": {
        "operationId": "RepairCaseController_listClientMessages",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "repairs"
        ]
      }
    },
    "/staff/repair-cases/scan/{rawQrToken}": {
      "get": {
        "operationId": "RepairCaseController_scanQrCode",
        "parameters": [
          {
            "name": "rawQrToken",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "repairs"
        ]
      }
    },
    "/staff/repair-cases/{id}/qr/regenerate": {
      "post": {
        "operationId": "RepairCaseController_regenerateQr",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "repairs"
        ]
      }
    },
    "/staff/repair-cases/{id}/qr/revoke": {
      "post": {
        "operationId": "RepairCaseController_revokeQr",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "repairs"
        ]
      }
    },
...
  },
  "info": {
    "title": "SaveMyControllers API",
    "description": "Contrat API du monolithe modulaire SaveMyControllers",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "smc_access_token"
      }
    },
...
  }
}